+++ semiPrj_BackPacker/src/main/java/com/bp/app/scheduler/controller/GetTimetable.java
@@ -1,4 +1,4 @@
-package com.bp.app.scheduler;
+package com.bp.app.scheduler.controller;
 
 import java.io.IOException;
 import java.io.PrintWriter;
@@ -12,42 +12,49 @@ import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
 
 import com.bp.app.scheduler.service.SchedulerService;
-import com.bp.app.scheduler.vo.SchedulerVo;
 import com.bp.app.scheduler.vo.TimetableVo;
 import com.google.gson.Gson;
 
-@WebServlet("/schedulermake/getScheduler")
-public class GetScheduler extends HttpServlet{
+@WebServlet("/schedulermake/getTimetable")
+public class GetTimetable extends HttpServlet{
 
-	
 	@Override
-	protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
+	protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
 
 		try {
+			//데이터 
 			
+			//서비스
 			SchedulerService ss = new SchedulerService();
-			List<SchedulerVo> sList = ss.getScheduler(req);
+			List<TimetableVo>list =  ss.getTimetable(req);
 			
 			//자바객체를 JSON 형태의 문자열로 변환
 			Gson gson = new Gson();
-			String sjsonStr = gson.toJson(sList);
+			String jsonStr = gson.toJson(list);
 			
+			System.out.println(list);
 			//문자열 내보내기
 			resp.setCharacterEncoding("UTF-8");
 			PrintWriter out = resp.getWriter();
-			out.write(sjsonStr);
-			System.out.println(sjsonStr);
+			out.write(jsonStr);
 			
 //			HttpSession session = req.getSession();
-//			session.setAttribute("scheduler", sjsonStr);
+//			session.setAttribute("timetable", jsonStr);
+//			System.out.println(jsonStr);
+			
+//			req.setAttribute("timetable", jsonStr);
+//			req.getRequestDispatcher("/WEB-INF/views/scheduler/makeScheduler.jsp").forward(req, resp);
+			
 			
 			
 			
 		} catch (Exception e) {
-			System.out.println("[ERROR] scheduler가져오다가 에러");
+			System.out.println("[ERROR] timetable가져오다가 에러");
 			e.printStackTrace();
 		}
-	
+		
 	
 	}
+	
+	
 }
