+++ semiPrj_BackPacker/src/main/java/com/bp/app/scheduler/controller/SchedulerMake.java
@@ -1,4 +1,4 @@
-package com.bp.app.scheduler;
+package com.bp.app.scheduler.controller;
 
 import java.io.IOException;
 import java.util.List;
